Move freedom from a dependency to a peer-dependency.
Node packages that depend on uproxy-lib do need freedom.js, but
their freedom.js may be provided by any one of several different
packages, including freedom, freedom-for-chrome, and freedom-for-firefox.
Since node doesn't have metapackages or disjunctive dependencies, the
best we can do is to specify freedom as a peer dependency (to avoid
version clashes) and a dev dependency (for the samples).