Remove some code duplication in logging
So, the original intention of this commit was to remove the code
duplication in the setConsoleFilter and the setBufferedLogFilter
functions, however, it kind of exploded from there.  It was originally
taken care of by just having them call a third function that would store
the result in the correct location, however, the switch statement looked
a bit silly.  That was resolved by by switching the filter arrays to
each be part of an indexed structure, however, it felt a bit silly
having the two checkFilter functions reference them in doRealLog.  This
was the best I could come up with, however, it does feel a bit
excessive.

It does not feel clean to me having the setConsoleFilter and
setBufferedLogFilter functions, however, I don't think that can be
streamlined anymore without changing the interface to the module.