Allow arguments to be passed to the logger functions without an array
This change feels a bit excessive, however, most of it was necessary to
get this done in a clean manner.  The objective here was to allow making
a log call to be done as `log('%1 %2', 'foo', 'bar')` instead of
`log('%1 %2', ['foo', 'bar'])`; not totally necessary but a bit nicer to
work with.

To do this, function signatures needed to be switched around for all
the main logging functions to accept variable-length arguments.  There
needed to be some code to ensure backwards compatibility and that felt
better in a private method rather than the doLog function.  The rest of
the code in doLog then felt redundant and the calling felt simplified by
moving it to a private method and giving it access to the logging
variable.

In terms of testing, the module has been broken up a bit more to have
separate tests for the tag-calling behaviour (which was previously mixed
into the tests) and then to have the tests for actually calling the
logging functions below that with a mockLogger object set up each time
so that did not have to be done in each test.  This provides better test
granuality and reduces the overhead for setting up new tests.