Handle non-string log elements and extra elements
When the log functions are called with anything that is not a string, it
is first json-ified before it is put into the log message.  This should
result in more readable output in the normal case.

If there are extra arguments to the log message that are not mentioned
in the original string, they are added to the end of the log message
instead of just being ignored.