Add the ability to set a minimum logging before messages get passed to freedom
One area we are facing a problem with performance-wise is passing
messages to freedom from the logger.  This involves sending the message
both to the core environment and then to the actual logging module.

To remove that overhead for messages we intend to do absolutely nothing
with, this adds the ability to set a minLogLevel on the the logger which
converts the message to a noop with no freedom overhead.

Adds a temporary solution for uProxy/uproxy#1005