Update local loggers with the minimum level to send logs
This makes a few changes to how the logging system works in an effort to
reduce the number of IPC calls we make for logging.

Background:

Right now, to user the logger in a project, you load up the local logger
(logger/logger.ts) in the context of your code and then load up the
logging provider and logging controller as freedom modules.  To actually
log a message, the logger will pass the data to be logged through an IPC
to the freedom module.  Previously, this was done for every single log
message.

In a lot of cases, we would end up then discarding the log message on
the remote side due to the logging level being set at a threshold above
the message that was just logged.  In this case, we used an expensive
IPC call for no reason.

Changes:

The big change here is the creation of a new freedom interface,
LoggingListener, that will monitor the minimum level at which any action
will be taken for log messages with a specific tag.  The logger will
automatically set itself up to listen for updates to its tag if the
loggingListener module is loaded, or otherwise will function as before.

In the LoggingController module, we register each listener that is
created to its tag and will send out updates whenever its minimum level
(the lowest level between the different logging destinations where a
message would be logged) is changed.

To clean up the code a bit, the functions to change the logging
destinations have changed slightly as well.  There are now three
different functions that can be used to affect the levels:

setDefaultFilter: changes the default logging level, this is used if the
                  level is not specified for a tag

setModuleFilters: set or clear the logging levels for a list of modules.
                  Modules should be passed in as an object with the key
                  being the module name and the value being the level it
                  should be set to.  Setting the level to default clears
                  any overrides for that module.

clearFilters:     Returns all modules to use the default level filters

All three functions take the first argument as which logging destination
(current console or buffered) the level should be changed for.  All
destinations and levels are specified in enums in
loggingprovider.types.ts.

Fixes uProxy/uproxy#1005