Add functions to the build rules that take care of building tests
This adds three new functions.

The first function (well, technically the last in the file...) takes
care of getting a list of all the tests there are to run.

The second function takes care of generating the browserify and jasmine
rules to build that single test and then run it in its own environment.

The third rule runs a single test in its own environment.

TODO: move to gulp